Clean architecture - UseCases. - Modularization. - Abstract data sources. - Data mapping and data models.
