+++ app/build.gradle.kts
@@ -63,6 +63,9 @@ android {
 
 dependencies {
 
+    implementation(project(":domain"))
+    implementation(project(":usecases"))
+    implementation(project(":data"))
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.lifecycle.runtime.ktx)
     implementation(libs.androidx.activity.compose)
