--- app/src/main/java/com/oscar/moviesproject/data/RegionRepository.kt
@@ -1,11 +1,0 @@
-package com.oscar.moviesproject.data
-
-import com.oscar.moviesproject.data.datasource.RegionDataSource
-
-
-class RegionRepository(
-    private val regionDataSource: RegionDataSource
-) {
-    suspend fun findLastRegion(): String = regionDataSource.findLastRegion()
-
-}
