--- app/src/main/java/com/oscar/moviesproject/data/datasource/LocationDataSource.kt
@@ -1,26 +1,0 @@
-package com.oscar.moviesproject.data.datasource
-
-import android.annotation.SuppressLint
-import android.app.Application
-import android.location.Location
-import com.google.android.gms.location.FusedLocationProviderClient
-import com.google.android.gms.location.LocationServices
-import kotlinx.coroutines.suspendCancellableCoroutine
-import kotlin.coroutines.resume
-
-class LocationDataSource(app: Application) {
-    private val fusedLocationClient = LocationServices.getFusedLocationProviderClient(app)
-
-    suspend fun findLastLocation(): Location? = fusedLocationClient.lastLocation()
-
-    @SuppressLint("MissingPermission")
-    suspend fun FusedLocationProviderClient.lastLocation() : Location?{
-        return suspendCancellableCoroutine { continuation ->
-            lastLocation.addOnSuccessListener { location ->
-                continuation.resume(null)
-            }.addOnFailureListener{
-                continuation.resume(null)
-            }
-        }
-    }
-}
