--- app/src/main/java/com/oscar/moviesproject/data/datasource/MoviesLocalDataSource.kt
@@ -1,16 +1,0 @@
-package com.oscar.moviesproject.data.datasource
-
-import com.oscar.moviesproject.data.Movie
-import com.oscar.moviesproject.data.datasource.database.MoviesDao
-
-class MoviesLocalDataSource(private val moviesDao: MoviesDao) {
-
-    val movies = moviesDao.fetchPopularMovies()
-
-    fun findMoviesById(id: Int) = moviesDao.findMovieById(id)
-
-    suspend fun isEmpty() = moviesDao.countMovies() == 0
-
-    suspend fun saveMovies(movies: List<Movie>) = moviesDao.saveMovies(movies)
-
-}
