--- app/src/main/java/com/oscar/moviesproject/data/datasource/MoviesRemoteDataSource.kt
@@ -1,36 +1,0 @@
-package com.oscar.moviesproject.data.datasource
-
-import com.oscar.moviesproject.data.Movie
-import com.oscar.moviesproject.data.datasource.remote.MoviesClient
-import com.oscar.moviesproject.data.datasource.remote.RemoteMovie
-
-class MoviesRemoteDataSource() {
-
-    suspend fun fetchPopularMovies(region: String): List<Movie> =
-        MoviesClient
-            .instance
-            .fetchPopularMovies(region)
-            .results
-            .map { it.toDomainModel() }
-
-    suspend fun findMovieById(id: Int): Movie =
-        MoviesClient
-            .instance
-            .fetchMovieById(id)
-            .toDomainModel()
-}
-
-private fun RemoteMovie.toDomainModel(): Movie =
-    Movie(
-        id = id,
-        title = title,
-        overview = overview,
-        releaseDate = releaseDate,
-        poster = "https://image.tmdb.org/t/p/w185/$posterPath",
-        backdrop = backdropPath.let { "https://image.tmdb.org/t/p/w780/$it" },
-        originalTitle = originalTitle,
-        originalLanguage = originalLanguage,
-        popularity = popularity,
-        voteAverage = voteAverage,
-        favorite = false
-    )
