+++ app/src/main/java/com/oscar/moviesproject/framework/database/DbMovie.kt
@@ -1,0 +1,20 @@
+package com.oscar.moviesproject.framework.database
+
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+
+@Entity
+data class DbMovie(
+    @PrimaryKey(autoGenerate = true)
+    val id: Int,
+    val title: String,
+    val overview: String,
+    val releaseDate: String,
+    val poster: String,
+    val backdrop: String?,
+    val originalTitle: String,
+    val originalLanguage: String,
+    val popularity: Double,
+    val voteAverage: Double,
+    val favorite: Boolean
+)
