+++ app/src/main/java/com/oscar/moviesproject/framework/remote/RemoteResult.kt
@@ -1,0 +1,26 @@
+package com.oscar.moviesproject.framework.remote
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class RemoteResult(
+    val page: Int,
+    val results: List<RemoteMovie>,
+    @SerialName("total_pages") val totalPages: Int,
+    @SerialName("total_results") val totalResults: Int
+)
+
+@Serializable
+data class RemoteMovie(
+    val id: Int,
+    val title: String,
+    val overview: String,
+    @SerialName("release_date") val releaseDate: String,
+    @SerialName("poster_path") val posterPath: String,
+    @SerialName("backdrop_path") val backdropPath: String,
+    @SerialName("original_title") val originalTitle: String,
+    @SerialName("original_language") val originalLanguage: String,
+    val popularity: Double,
+    @SerialName("vote_average") val voteAverage: Double
+)
