+++ app/src/main/java/com/oscar/moviesproject/ui/screens/Navigation.kt
@@ -1,6 +1,8 @@
 package com.oscar.moviesproject.ui.screens
 
+import android.location.Geocoder
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
 import androidx.compose.ui.platform.LocalContext
 import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavType
@@ -8,20 +10,19 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.navArgument
+import com.google.android.gms.location.LocationServices
 import com.oscar.moviesproject.App
-import com.oscar.moviesproject.data.MoviesRepository
-import com.oscar.moviesproject.data.RegionRepository
-import com.oscar.moviesproject.data.datasource.LocationDataSource
-import com.oscar.moviesproject.data.datasource.MoviesLocalDataSource
-import com.oscar.moviesproject.data.datasource.MoviesRemoteDataSource
-import com.oscar.moviesproject.data.datasource.RegionDataSource
+import com.oscar.data.MoviesRepository
+import com.oscar.data.RegionRepository
+import com.oscar.moviesproject.framework.GeocoderRegionDataSource
+import com.oscar.moviesproject.framework.MoviesRoomDataSource
+import com.oscar.moviesproject.framework.remote.MoviesClient
+import com.oscar.moviesproject.framework.MoviesServerDataSource
+import com.oscar.moviesproject.framework.PlayServicesLocationDataSource
 import com.oscar.moviesproject.ui.screens.detail.DetailScreen
 import com.oscar.moviesproject.ui.screens.detail.DetailViewModel
 import com.oscar.moviesproject.ui.screens.home.HomeScreen
 import com.oscar.moviesproject.ui.screens.home.HomeViewModel
-import com.oscar.moviesproject.usecases.FetchMoviesUseCase
-import com.oscar.moviesproject.usecases.FindMovieByIdUseCase
-import com.oscar.moviesproject.usecases.ToggleFavoriteUseCase
 
 sealed class NavScreen(val route: String) {
     data object Home : NavScreen(route = "home")
@@ -38,11 +39,22 @@ enum class NavArgs(val key: String) {
 fun Navigation() {
     val navController = rememberNavController()
     val app = LocalContext.current.applicationContext as App
-    val moviesRepository = MoviesRepository(
-        regionRepository = RegionRepository(RegionDataSource(app, LocationDataSource(app))),
-        localDataSource = MoviesLocalDataSource(app.db.moviesDao()),
-        remoteDataSource = MoviesRemoteDataSource()
-    )
+    val moviesRepository = remember {
+        com.oscar.data.MoviesRepository(
+            regionRepository = com.oscar.data.RegionRepository(
+                GeocoderRegionDataSource(
+                    geocoder = Geocoder(app),
+                    locationDataSource = PlayServicesLocationDataSource(
+                        LocationServices.getFusedLocationProviderClient(
+                            app
+                        )
+                    )
+                )
+            ),
+            localDataSource = MoviesRoomDataSource(app.db.moviesDao()),
+            remoteDataSource = MoviesServerDataSource(MoviesClient.instance)
+        )
+    }
     NavHost(navController = navController, startDestination = NavScreen.Home.route) {
         composable(NavScreen.Home.route) {
             HomeScreen(
@@ -51,7 +63,7 @@ fun Navigation() {
                         route = NavScreen.Detail.createRoute(movieId = movie.id)
                     )
                 },
-                viewModel { HomeViewModel(FetchMoviesUseCase(moviesRepository)) }
+                viewModel { HomeViewModel(com.oscar.usecases.FetchMoviesUseCase(moviesRepository)) }
             )
         }
         composable(
@@ -65,8 +77,8 @@ fun Navigation() {
                 vm = viewModel {
                     DetailViewModel(
                         movieId,
-                        ToggleFavoriteUseCase(moviesRepository),
-                        FindMovieByIdUseCase(moviesRepository)
+                        com.oscar.usecases.ToggleFavoriteUseCase(moviesRepository),
+                        com.oscar.usecases.FindMovieByIdUseCase(moviesRepository)
                     )
                 },
                 onBack = { navController.popBackStack() }
