+++ app/src/main/java/com/oscar/moviesproject/ui/screens/home/HomeViewModel.kt
@@ -3,9 +3,9 @@ package com.oscar.moviesproject.ui.screens.home
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.oscar.moviesproject.Result
-import com.oscar.moviesproject.data.Movie
+import com.oscar.domain.Movie
 import com.oscar.moviesproject.stateAsResultIn
-import com.oscar.moviesproject.usecases.FetchMoviesUseCase
+import com.oscar.usecases.FetchMoviesUseCase
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
@@ -13,7 +13,7 @@ import kotlinx.coroutines.flow.filter
 import kotlinx.coroutines.flow.flatMapLatest
 
 class HomeViewModel(
-    private val fetchMoviesUseCase: FetchMoviesUseCase
+    private val fetchMoviesUseCase: com.oscar.usecases.FetchMoviesUseCase
 ) : ViewModel() {
