--- app/src/main/java/com/oscar/moviesproject/usecases/ToggleFavoriteUseCase.kt
@@ -1,12 +1,0 @@
-package com.oscar.moviesproject.usecases
-
-import com.oscar.moviesproject.data.Movie
-import com.oscar.moviesproject.data.MoviesRepository
-
-class ToggleFavoriteUseCase(
-    private val moviesRepository: MoviesRepository
-) {
-
-    suspend operator fun invoke(movie: Movie) = moviesRepository.toggleFavorite(movie)
-
-}
