+++ data/src/main/java/com/oscar/data/RegionRepository.kt
@@ -1,0 +1,11 @@
+package com.oscar.data
+
+import com.oscar.data.datasource.RegionDataSource
+
+
+class RegionRepository(
+    private val regionDataSource: RegionDataSource
+) {
+    suspend fun findLastRegion(): String = regionDataSource.findLastRegion()
+
+}
