+++ data/src/main/java/com/oscar/data/datasource/LocationDataSource.kt
@@ -1,0 +1,8 @@
+package com.oscar.data.datasource
+
+import com.oscar.domain.Location
+
+
+interface LocationDataSource {
+    suspend fun findLastLocation(): Location?
+}
