+++ data/src/main/java/com/oscar/data/datasource/MoviesLocalDataSource.kt
@@ -1,0 +1,11 @@
+package com.oscar.data.datasource
+
+import com.oscar.domain.Movie
+import kotlinx.coroutines.flow.Flow
+
+interface MoviesLocalDataSource {
+    val movies: Flow<List<com.oscar.domain.Movie>>
+    fun findMoviesById(id: Int): Flow<com.oscar.domain.Movie?>
+    suspend fun isEmpty(): Boolean
+    suspend fun saveMovies(movies: List<com.oscar.domain.Movie>)
+}
