+++ data/src/main/java/com/oscar/data/datasource/MoviesRemoteDataSource.kt
@@ -1,0 +1,9 @@
+package com.oscar.data.datasource
+
+import com.oscar.domain.Movie
+
+interface MoviesRemoteDataSource {
+    suspend fun fetchPopularMovies(region: String): List<com.oscar.domain.Movie>
+
+    suspend fun findMovieById(id: Int): com.oscar.domain.Movie
+}
