+++ data/src/main/java/com/oscar/data/datasource/RegionDataSource.kt
@@ -1,0 +1,9 @@
+package com.oscar.data.datasource
+
+import com.oscar.domain.Location
+
+
+interface RegionDataSource {
+    suspend fun findLastRegion(): String
+    suspend fun Location.toRegion(): String
+}
