+++ domain/src/main/java/com/oscar/domain/Location.kt
@@ -1,0 +1,4 @@
+package com.oscar.domain
+
+class Location(val latitude: Double, val longitude: Double) {
+}
