+++ usecases/build.gradle.kts
@@ -1,0 +1,15 @@
+plugins {
+    id("java-library")
+    alias(libs.plugins.jetbrains.kotlin.jvm)
+}
+
+java {
+    sourceCompatibility = JavaVersion.VERSION_17
+    targetCompatibility = JavaVersion.VERSION_17
+}
+
+dependencies {
+    implementation(project(":data"))
+    implementation(project(":domain"))
+    implementation(libs.kotlinx.coroutines.core)
+}
