+++ usecases/src/main/java/com/oscar/usecases/ToggleFavoriteUseCase.kt
@@ -1,0 +1,12 @@
+package com.oscar.usecases
+
+import com.oscar.data.MoviesRepository
+import com.oscar.domain.Movie
+
+class ToggleFavoriteUseCase(
+    private val moviesRepository: MoviesRepository
+) {
+
+    suspend operator fun invoke(movie: Movie) = moviesRepository.toggleFavorite(movie)
+
+}
